test: add failing assert for bug in tric.h
When skipping a test with SKIP_TEST the test results of the following
tests are incorrect. The reason is that the function
tric_skip_test_execution() of tric.h does not set the mode of the
context to MODE_RESET. Add a failing assert to test
test_skip_test_execution_ok in test/tric_test.c to mark the bug.